requirements.txt
mysql-connector-python==8.0.29
pyfiglet==0.8.post1
